MP1/requirements.txt
gym
absl-py
jupyter
seaborn
matplotlib
Pillow
tqdm
scikit-learn
torch
torchvision
unidecode
